+++ src/routes/productsRoutes.js
@@ -1,55 +1,12 @@
 const express = require("express");
-const path = require("path");
 const router = express.Router();
 
-const multer = require("multer");
-const { body } = require("express-validator");
-
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, path.join(__dirname, "../../public/images/covers"));
-  },
-  filename: function (req, file, cb) {
-    cb(null, `${Date.now()}_img_${path.extname(file.originalname)}`);
-  },
-});
-
-const uploadFile = multer({ storage });
-
+//Controller
 const productsController = require("../controllers/productsController");
 
-const validations = [
-  body("title").notEmpty().withMessage("El nombre del álbum es requerido"),
-  body("artist").notEmpty().withMessage("El nombre del artista es requerido"),
-  body("description")
-    .notEmpty()
-    .withMessage("La descripción del disco es requerida"),
-  body("image").custom((value, { req }) => {
-    let file = req.file;
-    let acceptedExtensions = [".jpg", ".png"];
-
-    if (!file) {
-      throw new Error("Tienes que subir una imagen");
-    } else {
-      let fileExtension = path.extname(file.originalname);
-
-      if (!acceptedExtensions.includes(fileExtension)) {
-        throw new Error(
-          `Las extensiones permitidas son ${acceptedExtensions.join(",")}`
-        );
-      }
-    }
-
-    return true;
-  }),
-  body("tracklist")
-    .notEmpty()
-    .withMessage("Los tracks del disco son requeridos"),
-  body("genre").notEmpty().withMessage("Elige un género"),
-  body("release_date").notEmpty().withMessage("Escribe un año"),
-  body("price").notEmpty().withMessage("Escribe un precio"),
-  body("id_categories").notEmpty().withMessage("Elige una categoría"),
-];
+//Middlewares
+const validations = require("../middlewares/validateAlbumMiddleware");
+const uploadFile = require("../middlewares/multerAlbumCoverMiddleware");
 
 //Estas son subrutas de la ruta /products que se configura en app
