+++ src/routes/usersRoutes.js
@@ -1,5 +1,4 @@
 const express = require("express");
-
 const router = express.Router();
 
 //Controller
